+++ src/Service/Models/ProofOfWorkDto.cs
@@ -4,11 +4,11 @@ public record ProofOfWorkDto
 { 
     public int SecondsUntilBlock { get; init; }
 
-    public long TotalHashRate { get; set; } 
-
+    public double TotalHashRate { get; set; } 
+    
     public double Difficulty { get; set; }
-
+    
     public double Expected { get; set; } 
-
+    
     public string Threshold { get; set; } = "";
 }
